requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.0.1
sqlparse==0.4.2
